Clean up matchmaking index
