--- src/game-server/.env.development
@@ -1,1 +1,0 @@
-SAMPLE=development
