--- src/game-server/.env.production
@@ -1,1 +1,0 @@
-SAMPLE=production
