+++ src/matchmaking/.env.example
@@ -1,14 +1,9 @@
 # Server Configuration
 HOST=0.0.0.0
 PORT=3005
-NODE_ENV=development
-LOG_LEVEL=info
 
 # Database Configuration
 DATABASE_URL=file:./prisma/dev.db
-
-# CORS Configuration
-ALLOWED_ORIGINS=http://localhost:3000,http://localhost:8080
 
 # Matchmaking Configuration
 # Auto-pairing interval in milliseconds (default: 5000 = 5 seconds)
