+++ src/matchmaking/package.json
@@ -18,8 +18,6 @@
   "author": "Vova Shkonda",
   "license": "ISC",
   "dependencies": {
-    "@fastify/cors": "^9.0.1",
-    "@fastify/helmet": "^11.1.1",
     "@prisma/adapter-better-sqlite3": "^7.2.0",
     "@prisma/client": "^7.2.0",
     "async-mutex": "^0.5.0",
